+++ osdep/mac/libmpv_helper.swift
