+++ osdep/mac/log_helper.swift
