+++ osdep/mac/precise_timer.swift
