+++ osdep/mac/swift_bridge.h
@@ -30,8 +30,8 @@
 #include "input/event.h"
 #include "video/out/win_state.h"
 
-#include "osdep/macosx_application_objc.h"
-#include "osdep/macosx_events_objc.h"
+#include "osdep/mac/application_objc.h"
+#include "osdep/mac/events_objc.h"
 
 
 // complex macros won't get imported to Swift so we have to reassign them
