+++ osdep/mac/swift_compat.swift
