+++ osdep/mac/swift_extensions.swift
