+++ osdep/main-fn-mac.c
@@ -1,4 +1,4 @@
-#include "osdep/macosx_application.h"
+#include "osdep/mac/application.h"
 
 // This is needed because Cocoa absolutely requires creating the NSApplication
 // singleton and running it in the "main" thread. It is apparently not
