+++ osdep/path.h
@@ -26,7 +26,7 @@ typedef const char *(*mp_get_platform_path_cb)(void *talloc_ctx, const char *typ
 const char *mp_get_platform_path_darwin(void *talloc_ctx, const char *type);
 const char *mp_get_platform_path_uwp(void *talloc_ctx, const char *type);
 const char *mp_get_platform_path_win(void *talloc_ctx, const char *type);
-const char *mp_get_platform_path_osx(void *talloc_ctx, const char *type);
+const char *mp_get_platform_path_mac(void *talloc_ctx, const char *type);
 const char *mp_get_platform_path_unix(void *talloc_ctx, const char *type);
 
 #endif
