+++ osdep/semaphore-mac.c
